+++ packages/guide/src/app/locations/components/ranch/ranch.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 import { BaseShopComponent } from "../base-shop/base-shop.component";
-import { ShopItemData, ShopName } from "@ci/data-types";
+import { ProductSizeByMood, ShopItemData, ShopName } from "@ci/data-types";
 import { combineLatest, Observable, of, switchMap } from "rxjs";
 import { MappedAnimalShopData } from "../../types/mapped-animal-shop-data.type";
 
@@ -14,11 +14,13 @@ export class RanchComponent extends BaseShopComponent {
     protected animalData$: Observable<MappedAnimalShopData[]>;
 
     protected selectedAnimal?: MappedAnimalShopData
+    protected sizeByMood$: Observable<ProductSizeByMood[]>;
 
     constructor() {
         super();
         this.shopItemData$ = this._database.fetchShopItemData$(this.shopName);
         this.openingHours$ = this._database.fetchOpeningHours$(this.shopName);
+        this.sizeByMood$ = this._database.fetchAnimalMoodData$();
         this.animalData$ = combineLatest([
             this._database.fetchAnimals$(),
             this._database.fetchAnimalShopData$()